pyproject.toml
[tool.poetry]
name = "event-app"
version = "0.1.0"
description = ""
authors = ["MD Nurujjaman Nur <nur.ruet12@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
django-filter = "^22.1"

[tool.poetry.group.dev.dependencies]
ipython = "^8.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
